Refactoring
[tfs-changeset: 1651643]
