+++ src/System.Private.TypeLoader/src/Internal/Runtime/CompilerHelpers/LibraryInitializer.cs
@@ -13,8 +13,7 @@ namespace Internal.Runtime.CompilerHelpers
         public static void InitializeLibrary()
         {
             TypeLoaderEnvironment.Initialize();
-            // TODO: Find a better home for the FuncEval runtime support
-            SerializedDebugData.Initialize();
+            DebugFuncEval.Initialize();
         }
     }
 }
